reflets-transforms/requirements.txt
maltego-trx
requests~=2.27.1
pyyaml